requirements.txt
Flask>=2.0
Flask-Login>=0.6
Flask-SQLAlchemy>=2.5
pandas>=1.0
numpy>=1.18
pyserial>=3.4
gunicorn>=20.0
python-dotenv>=0.20